requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
identify==2.2.2
idna==2.10
importlib-metadata==3.10.0
lxml==4.6.3
nodeenv==1.5.0
packaging==20.9
pluggy==0.13.1
pre-commit==2.11.1
py==1.10.0
pyparsing==2.4.7
python-telegram-bot==13.4.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
tox==3.23.0
translate==3.5.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.4.3
zipp==3.4.1
